jtag/drivers/bcm2835gpio: Support GPIOs >= 32
Previously, not all 54 pins were supported by the driver.

Performance was cited as being the primary justification, notably:

  1. GPIO values cannot be written in bulk
  2. There is overhead for calculating the memory offset for the pin

In order to support pins >=32, the memory offset has to be calculated
for each pin in use in order to be toggled. There's no guarantee that
the mix of pins being used are backed by the same memory address, so
bulk operations cannot be performed when all pins are supported.

Add a new configure option to expand usable GPIO pins for the bcm2835
jtag driver. This drives a new define which will conditionally compile
code to support pins >32. This is disabled by default and previous code
patterns, including bulk GPIO operations, are used for the default case.
When the config option is enabled, bulk operations are disabled and
read/set/write ops are used for each pin being toggled.

The performance cost for calculating the pin offset can largely be
mitigated by using `unsigned` values so that integer division and modulo
operations can be accomplished with cheap bit shift math.

Update macros to provide consistent access to all pins, handle casting,
and to conform to style recommendations. New bulk macros are provided to
preserve behavior for builds that only use the first 32 pins.

Add support for changing slew rate and hysteresis for additional pads.

Change-Id: I18853d1a2c86776658630326c71a6bf236fcc6da
Signed-off-by: Jacob Zarnstorff <jacobzarnstorff@gmail.com>
[vfazio: add configure option, conditional compile statements, and casts]
Signed-off-by: Vincent Fazio <vfazio@gmail.com>